refactor(tests): remove unnecessary comments and improve test clarity across various plugin test files
