+++ web/app/components/plugins/install-plugin/base/loading.spec.tsx
@@ -2,12 +2,6 @@ import { render, screen } from '@testing-library/react'
 import * as React from 'react'
 import { describe, expect, it, vi } from 'vitest'
 
-vi.mock('@/app/components/base/checkbox', () => ({
-  default: ({ checked, disabled }: { checked: boolean, disabled: boolean }) => (
-    <input type="checkbox" data-testid="checkbox" checked={checked} disabled={disabled} readOnly />
-  ),
-}))
-
 vi.mock('../../card/base/placeholder', () => ({
   default: () => <div data-testid="placeholder" />,
 }))
@@ -24,9 +18,7 @@ describe('Loading', () => {
   it('should render disabled unchecked checkbox', () => {
     render(<Loading />)
 
-    const checkbox = screen.getByTestId('checkbox')
-    expect(checkbox).toBeDisabled()
-    expect(checkbox).not.toBeChecked()
+    expect(screen.getByTestId('checkbox-undefined')).toBeInTheDocument()
   })
 
   it('should render placeholder', () => {
