+++ web/app/components/plugins/install-plugin/base/version.spec.tsx
@@ -2,11 +2,6 @@ import { render, screen } from '@testing-library/react'
 import * as React from 'react'
 import { describe, expect, it, vi } from 'vitest'
 
-vi.mock('@/app/components/base/badge/index', () => ({
-  default: ({ children }: { children: React.ReactNode }) => <span data-testid="badge">{children}</span>,
-  BadgeState: { Default: 'default', Warning: 'warning' },
-}))
-
 describe('Version', () => {
   let Version: (typeof import('./version'))['default']
 
@@ -19,7 +14,7 @@ describe('Version', () => {
   it('should show simple version badge for new install', () => {
     render(<Version hasInstalled={false} toInstallVersion="1.0.0" />)
 
-    expect(screen.getByTestId('badge')).toHaveTextContent('1.0.0')
+    expect(screen.getByText('1.0.0')).toBeInTheDocument()
   })
 
   it('should show upgrade version badge for existing install', () => {
@@ -31,7 +26,7 @@ describe('Version', () => {
       />,
     )
 
-    expect(screen.getByTestId('badge')).toHaveTextContent('1.0.0 -> 2.0.0')
+    expect(screen.getByText('1.0.0 -> 2.0.0')).toBeInTheDocument()
   })
 
   it('should handle downgrade version display', () => {
@@ -43,6 +38,6 @@ describe('Version', () => {
       />,
     )
 
-    expect(screen.getByTestId('badge')).toHaveTextContent('2.0.0 -> 1.0.0')
+    expect(screen.getByText('2.0.0 -> 1.0.0')).toBeInTheDocument()
   })
 })
