+++ web/app/components/plugins/marketplace/index.spec.tsx
@@ -1,1828 +1,15 @@
-import type { MarketplaceCollection } from './types'
-import type { Plugin } from '@/app/components/plugins/types'
-import { act, render, renderHook } from '@testing-library/react'
-import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
-import { PluginCategoryEnum } from '@/app/components/plugins/types'
-
-// ================================
-// Import Components After Mocks
-// ================================
-
-// Note: Import after mocks are set up
-import { DEFAULT_SORT, PLUGIN_TYPE_SEARCH_MAP, SCROLL_BOTTOM_THRESHOLD } from './constants'
-import {
-  getFormattedPlugin,
-  getMarketplaceListCondition,
-  getMarketplaceListFilterType,
-  getPluginDetailLinkInMarketplace,
-  getPluginIconInMarketplace,
-  getPluginLinkInMarketplace,
-} from './utils'
-
-// ================================
-// Mock External Dependencies Only
-// ================================
-
-// Mock i18next-config
-vi.mock('@/i18n-config/i18next-config', () => ({
-  default: {
-    getFixedT: (_locale: string) => (key: string, options?: Record<string, unknown>) => {
-      if (options && options.ns) {
-        return `${options.ns}.${key}`
-      }
-      else {
-        return key
-      }
-    },
-  },
-}))
-
-// Mock use-query-params hook
-const mockSetUrlFilters = vi.fn()
-vi.mock('@/hooks/use-query-params', () => ({
-  useMarketplaceFilters: () => [
-    { q: '', tags: [], category: '' },
-    mockSetUrlFilters,
-  ],
-}))
-
-// Mock use-plugins service
-const mockInstalledPluginListData = {
-  plugins: [],
-}
-vi.mock('@/service/use-plugins', () => ({
-  useInstalledPluginList: (_enabled: boolean) => ({
-    data: mockInstalledPluginListData,
-    isSuccess: true,
-  }),
-}))
-
-// Mock tanstack query
-const mockFetchNextPage = vi.fn()
-const mockHasNextPage = false
-let mockInfiniteQueryData: { pages: Array<{ plugins: unknown[], total: number, page: number, page_size: number }> } | undefined
-let capturedInfiniteQueryFn: ((ctx: { pageParam: number, signal: AbortSignal }) => Promise<unknown>) | null = null
-let capturedQueryFn: ((ctx: { signal: AbortSignal }) => Promise<unknown>) | null = null
-let capturedGetNextPageParam: ((lastPage: { page: number, page_size: number, total: number }) => number | undefined) | null = null
-
-vi.mock('@tanstack/react-query', () => ({
-  useQuery: vi.fn(({ queryFn, enabled }: { queryFn: (ctx: { signal: AbortSignal }) => Promise<unknown>, enabled: boolean }) => {
-    // Capture queryFn for later testing
-    capturedQueryFn = queryFn
-    // Always call queryFn to increase coverage (including when enabled is false)
-    if (queryFn) {
-      const controller = new AbortController()
-      queryFn({ signal: controller.signal }).catch(() => {})
-    }
-    return {
-      data: enabled ? { marketplaceCollections: [], marketplaceCollectionPluginsMap: {} } : undefined,
-      isFetching: false,
-      isPending: false,
-      isSuccess: enabled,
-    }
-  }),
-  useInfiniteQuery: vi.fn(({ queryFn, getNextPageParam, enabled: _enabled }: {
-    queryFn: (ctx: { pageParam: number, signal: AbortSignal }) => Promise<unknown>
-    getNextPageParam: (lastPage: { page: number, page_size: number, total: number }) => number | undefined
-    enabled: boolean
-  }) => {
-    // Capture queryFn and getNextPageParam for later testing
-    capturedInfiniteQueryFn = queryFn
-    capturedGetNextPageParam = getNextPageParam
-    // Always call queryFn to increase coverage (including when enabled is false for edge cases)
-    if (queryFn) {
-      const controller = new AbortController()
-      queryFn({ pageParam: 1, signal: controller.signal }).catch(() => {})
-    }
-    // Call getNextPageParam to increase coverage
-    if (getNextPageParam) {
-      // Test with more data available
-      getNextPageParam({ page: 1, page_size: 40, total: 100 })
-      // Test with no more data
-      getNextPageParam({ page: 3, page_size: 40, total: 100 })
-    }
-    return {
-      data: mockInfiniteQueryData,
-      isPending: false,
-      isFetching: false,
-      isFetchingNextPage: false,
-      hasNextPage: mockHasNextPage,
-      fetchNextPage: mockFetchNextPage,
-    }
-  }),
-  useQueryClient: vi.fn(() => ({
-    removeQueries: vi.fn(),
-  })),
-}))
-
-// Mock ahooks
-vi.mock('ahooks', () => ({
-  useDebounceFn: (fn: (...args: unknown[]) => void) => ({
-    run: fn,
-    cancel: vi.fn(),
-  }),
-}))
-
-// Mock marketplace service
-let mockPostMarketplaceShouldFail = false
-const mockPostMarketplaceResponse: {
-  data: {
-    plugins: Array<{ type: string, org: string, name: string, tags: unknown[] }>
-    bundles: Array<{ type: string, org: string, name: string, tags: unknown[] }>
-    total: number
-  }
-} = {
-  data: {
-    plugins: [
-      { type: 'plugin', org: 'test', name: 'plugin1', tags: [] },
-      { type: 'plugin', org: 'test', name: 'plugin2', tags: [] },
-    ],
-    bundles: [],
-    total: 2,
-  },
-}
-vi.mock('@/service/base', () => ({
-  postMarketplace: vi.fn(() => {
-    if (mockPostMarketplaceShouldFail)
-      return Promise.reject(new Error('Mock API error'))
-    return Promise.resolve(mockPostMarketplaceResponse)
-  }),
-}))
-
-// Mock config
-vi.mock('@/config', () => ({
-  API_PREFIX: '/api',
-  APP_VERSION: '1.0.0',
-  IS_MARKETPLACE: false,
-  MARKETPLACE_API_PREFIX: 'https://marketplace.dify.ai/api/v1',
-}))
-
-// Mock var utils
-vi.mock('@/utils/var', () => ({
-  getMarketplaceUrl: (path: string, _params?: Record<string, string | undefined>) => `https://marketplace.dify.ai${path}`,
-}))
-
-// Mock marketplace client used by marketplace utils
-vi.mock('@/service/client', () => ({
-  marketplaceClient: {
-    collections: vi.fn(async (_args?: unknown, _opts?: { signal?: AbortSignal }) => ({
-      data: {
-        collections: [
-          {
-            name: 'collection-1',
-            label: { 'en-US': 'Collection 1' },
-            description: { 'en-US': 'Desc' },
-            rule: '',
-            created_at: '2024-01-01',
-            updated_at: '2024-01-01',
-            searchable: true,
-            search_params: { query: '', sort_by: 'install_count', sort_order: 'DESC' },
-          },
-        ],
-      },
-    })),
-    collectionPlugins: vi.fn(async (_args?: unknown, _opts?: { signal?: AbortSignal }) => ({
-      data: {
-        plugins: [
-          { type: 'plugin', org: 'test', name: 'plugin1', tags: [] },
-        ],
-      },
-    })),
-    // Some utils paths may call searchAdvanced; provide a minimal stub
-    searchAdvanced: vi.fn(async (_args?: unknown, _opts?: { signal?: AbortSignal }) => ({
-      data: {
-        plugins: [
-          { type: 'plugin', org: 'test', name: 'plugin1', tags: [] },
-        ],
-        total: 1,
-      },
-    })),
-  },
-}))
-
-// Mock context/query-client
-vi.mock('@/context/query-client', () => ({
-  TanstackQueryInitializer: ({ children }: { children: React.ReactNode }) => <div data-testid="query-initializer">{children}</div>,
-}))
-
-// Mock i18n-config/server
-vi.mock('@/i18n-config/server', () => ({
-  getLocaleOnServer: vi.fn(() => Promise.resolve('en-US')),
-  getTranslation: vi.fn(() => Promise.resolve({ t: (key: string) => key })),
-}))
-
-// Mock useTheme hook
-const mockTheme = 'light'
-vi.mock('@/hooks/use-theme', () => ({
-  default: () => ({
-    theme: mockTheme,
-  }),
-}))
-
-// Mock next-themes
-vi.mock('next-themes', () => ({
-  useTheme: () => ({
-    theme: mockTheme,
-  }),
-}))
-
-// Mock useLocale context
-vi.mock('@/context/i18n', () => ({
-  useLocale: () => 'en-US',
-}))
-
-// Mock i18n-config/language
-vi.mock('@/i18n-config/language', () => ({
-  getLanguage: (locale: string) => locale || 'en-US',
-}))
-
-// Mock global fetch for utils testing
-const originalFetch = globalThis.fetch
-
-// Mock useTags hook
-const mockTags = [
-  { name: 'search', label: 'Search' },
-  { name: 'image', label: 'Image' },
-  { name: 'agent', label: 'Agent' },
-]
-
-const mockTagsMap = mockTags.reduce((acc, tag) => {
-  acc[tag.name] = tag
-  return acc
-}, {} as Record<string, { name: string, label: string }>)
-
-vi.mock('@/app/components/plugins/hooks', () => ({
-  useTags: () => ({
-    tags: mockTags,
-    tagsMap: mockTagsMap,
-    getTagLabel: (name: string) => {
-      const tag = mockTags.find(t => t.name === name)
-      return tag?.label || name
-    },
-  }),
-}))
-
-// Mock plugins utils
-vi.mock('../utils', () => ({
-  getValidCategoryKeys: (category: string | undefined) => category || '',
-  getValidTagKeys: (tags: string[] | string | undefined) => {
-    if (Array.isArray(tags))
-      return tags
-    if (typeof tags === 'string')
-      return tags.split(',').filter(Boolean)
-    return []
-  },
-}))
-
-// Mock portal-to-follow-elem with shared open state
-let mockPortalOpenState = false
-
-vi.mock('@/app/components/base/portal-to-follow-elem', () => ({
-  PortalToFollowElem: ({ children, open }: {
-    children: React.ReactNode
-    open: boolean
-  }) => {
-    mockPortalOpenState = open
-    return (
-      <div data-testid="portal-elem" data-open={open}>
-        {children}
-      </div>
-    )
-  },
-  PortalToFollowElemTrigger: ({ children, onClick, className }: {
-    children: React.ReactNode
-    onClick: () => void
-    className?: string
-  }) => (
-    <div data-testid="portal-trigger" onClick={onClick} className={className}>
-      {children}
-    </div>
-  ),
-  PortalToFollowElemContent: ({ children, className }: {
-    children: React.ReactNode
-    className?: string
-  }) => {
-    if (!mockPortalOpenState)
-      return null
-    return (
-      <div data-testid="portal-content" className={className}>
-        {children}
-      </div>
-    )
-  },
-}))
-
-// Mock Card component
-vi.mock('@/app/components/plugins/card', () => ({
-  default: ({ payload, footer }: { payload: Plugin, footer?: React.ReactNode }) => (
-    <div data-testid={`card-${payload.name}`}>
-      <div data-testid="card-name">{payload.name}</div>
-      {!!footer && <div data-testid="card-footer">{footer}</div>}
-    </div>
-  ),
-}))
-
-// Mock CardMoreInfo component
-vi.mock('@/app/components/plugins/card/card-more-info', () => ({
-  default: ({ downloadCount, tags }: { downloadCount: number, tags: string[] }) => (
-    <div data-testid="card-more-info">
-      <span data-testid="download-count">{downloadCount}</span>
-      <span data-testid="tags">{tags.join(',')}</span>
-    </div>
-  ),
-}))
-
-// Mock InstallFromMarketplace component
-vi.mock('@/app/components/plugins/install-plugin/install-from-marketplace', () => ({
-  default: ({ onClose }: { onClose: () => void }) => (
-    <div data-testid="install-from-marketplace">
-      <button onClick={onClose} data-testid="close-install-modal">Close</button>
-    </div>
-  ),
-}))
-
-// Mock base icons
-vi.mock('@/app/components/base/icons/src/vender/other', () => ({
-  Group: ({ className }: { className?: string }) => <span data-testid="group-icon" className={className} />,
-}))
-
-vi.mock('@/app/components/base/icons/src/vender/plugin', () => ({
-  Trigger: ({ className }: { className?: string }) => <span data-testid="trigger-icon" className={className} />,
-}))
-
-// ================================
-// Test Data Factories
-// ================================
-
-const createMockPlugin = (overrides?: Partial<Plugin>): Plugin => ({
-  type: 'plugin',
-  org: 'test-org',
-  name: `test-plugin-${Math.random().toString(36).substring(7)}`,
-  plugin_id: `plugin-${Math.random().toString(36).substring(7)}`,
-  version: '1.0.0',
-  latest_version: '1.0.0',
-  latest_package_identifier: 'test-org/test-plugin:1.0.0',
-  icon: '/icon.png',
-  verified: true,
-  label: { 'en-US': 'Test Plugin' },
-  brief: { 'en-US': 'Test plugin brief description' },
-  description: { 'en-US': 'Test plugin full description' },
-  introduction: 'Test plugin introduction',
-  repository: 'https://github.com/test/plugin',
-  category: PluginCategoryEnum.tool,
-  install_count: 1000,
-  endpoint: { settings: [] },
-  tags: [{ name: 'search' }],
-  badges: [],
-  verification: { authorized_category: 'community' },
-  from: 'marketplace',
-  ...overrides,
-})
-
-const createMockPluginList = (count: number): Plugin[] =>
-  Array.from({ length: count }, (_, i) =>
-    createMockPlugin({
-      name: `plugin-${i}`,
-      plugin_id: `plugin-id-${i}`,
-      install_count: 1000 - i * 10,
-    }))
-
-const createMockCollection = (overrides?: Partial<MarketplaceCollection>): MarketplaceCollection => ({
-  name: 'test-collection',
-  label: { 'en-US': 'Test Collection' },
-  description: { 'en-US': 'Test collection description' },
-  rule: 'test-rule',
-  created_at: '2024-01-01',
-  updated_at: '2024-01-01',
-  searchable: true,
-  search_params: {
-    query: '',
-    sort_by: 'install_count',
-    sort_order: 'DESC',
-  },
-  ...overrides,
-})
-
-// ================================
-// Constants Tests
-// ================================
-describe('constants', () => {
-  describe('DEFAULT_SORT', () => {
-    it('should have correct default sort values', () => {
-      expect(DEFAULT_SORT).toEqual({
-        sortBy: 'install_count',
-        sortOrder: 'DESC',
-      })
-    })
-
-    it('should be immutable at runtime', () => {
-      const originalSortBy = DEFAULT_SORT.sortBy
-      const originalSortOrder = DEFAULT_SORT.sortOrder
-
-      expect(DEFAULT_SORT.sortBy).toBe(originalSortBy)
-      expect(DEFAULT_SORT.sortOrder).toBe(originalSortOrder)
-    })
-  })
-
-  describe('SCROLL_BOTTOM_THRESHOLD', () => {
-    it('should be 100 pixels', () => {
-      expect(SCROLL_BOTTOM_THRESHOLD).toBe(100)
-    })
-  })
-})
-
-// ================================
-// PLUGIN_TYPE_SEARCH_MAP Tests
-// ================================
-describe('PLUGIN_TYPE_SEARCH_MAP', () => {
-  it('should contain all expected keys', () => {
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('all')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('model')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('tool')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('agent')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('extension')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('datasource')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('trigger')
-    expect(PLUGIN_TYPE_SEARCH_MAP).toHaveProperty('bundle')
-  })
-
-  it('should map to correct category enum values', () => {
-    expect(PLUGIN_TYPE_SEARCH_MAP.all).toBe('all')
-    expect(PLUGIN_TYPE_SEARCH_MAP.model).toBe(PluginCategoryEnum.model)
-    expect(PLUGIN_TYPE_SEARCH_MAP.tool).toBe(PluginCategoryEnum.tool)
-    expect(PLUGIN_TYPE_SEARCH_MAP.agent).toBe(PluginCategoryEnum.agent)
-    expect(PLUGIN_TYPE_SEARCH_MAP.extension).toBe(PluginCategoryEnum.extension)
-    expect(PLUGIN_TYPE_SEARCH_MAP.datasource).toBe(PluginCategoryEnum.datasource)
-    expect(PLUGIN_TYPE_SEARCH_MAP.trigger).toBe(PluginCategoryEnum.trigger)
-    expect(PLUGIN_TYPE_SEARCH_MAP.bundle).toBe('bundle')
-  })
-})
-
-// ================================
-// Utils Tests
-// ================================
-describe('utils', () => {
-  describe('getPluginIconInMarketplace', () => {
-    it('should return correct icon URL for regular plugin', () => {
-      const plugin = createMockPlugin({ org: 'test-org', name: 'test-plugin', type: 'plugin' })
-      const iconUrl = getPluginIconInMarketplace(plugin)
-
-      expect(iconUrl).toBe('https://marketplace.dify.ai/api/v1/plugins/test-org/test-plugin/icon')
-    })
-
-    it('should return correct icon URL for bundle', () => {
-      const bundle = createMockPlugin({ org: 'test-org', name: 'test-bundle', type: 'bundle' })
-      const iconUrl = getPluginIconInMarketplace(bundle)
-
-      expect(iconUrl).toBe('https://marketplace.dify.ai/api/v1/bundles/test-org/test-bundle/icon')
-    })
-  })
-
-  describe('getFormattedPlugin', () => {
-    it('should format plugin with icon URL', () => {
-      const rawPlugin = {
-        type: 'plugin',
-        org: 'test-org',
-        name: 'test-plugin',
-        tags: [{ name: 'search' }],
-      } as unknown as Plugin
-
-      const formatted = getFormattedPlugin(rawPlugin)
-
-      expect(formatted.icon).toBe('https://marketplace.dify.ai/api/v1/plugins/test-org/test-plugin/icon')
-    })
-
-    it('should format bundle with additional properties', () => {
-      const rawBundle = {
-        type: 'bundle',
-        org: 'test-org',
-        name: 'test-bundle',
-        description: 'Bundle description',
-        labels: { 'en-US': 'Test Bundle' },
-      } as unknown as Plugin
-
-      const formatted = getFormattedPlugin(rawBundle)
-
-      expect(formatted.icon).toBe('https://marketplace.dify.ai/api/v1/bundles/test-org/test-bundle/icon')
-      expect(formatted.brief).toBe('Bundle description')
-      expect(formatted.label).toEqual({ 'en-US': 'Test Bundle' })
-    })
-  })
-
-  describe('getPluginLinkInMarketplace', () => {
-    it('should return correct link for regular plugin', () => {
-      const plugin = createMockPlugin({ org: 'test-org', name: 'test-plugin', type: 'plugin' })
-      const link = getPluginLinkInMarketplace(plugin)
-
-      expect(link).toBe('https://marketplace.dify.ai/plugins/test-org/test-plugin')
-    })
-
-    it('should return correct link for bundle', () => {
-      const bundle = createMockPlugin({ org: 'test-org', name: 'test-bundle', type: 'bundle' })
-      const link = getPluginLinkInMarketplace(bundle)
-
-      expect(link).toBe('https://marketplace.dify.ai/bundles/test-org/test-bundle')
-    })
-  })
-
-  describe('getPluginDetailLinkInMarketplace', () => {
-    it('should return correct detail link for regular plugin', () => {
-      const plugin = createMockPlugin({ org: 'test-org', name: 'test-plugin', type: 'plugin' })
-      const link = getPluginDetailLinkInMarketplace(plugin)
-
-      expect(link).toBe('/plugins/test-org/test-plugin')
-    })
-
-    it('should return correct detail link for bundle', () => {
-      const bundle = createMockPlugin({ org: 'test-org', name: 'test-bundle', type: 'bundle' })
-      const link = getPluginDetailLinkInMarketplace(bundle)
-
-      expect(link).toBe('/bundles/test-org/test-bundle')
-    })
-  })
-
-  describe('getMarketplaceListCondition', () => {
-    it('should return category condition for tool', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.tool)).toBe('category=tool')
-    })
-
-    it('should return category condition for model', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.model)).toBe('category=model')
-    })
-
-    it('should return category condition for agent', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.agent)).toBe('category=agent-strategy')
-    })
-
-    it('should return category condition for datasource', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.datasource)).toBe('category=datasource')
-    })
-
-    it('should return category condition for trigger', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.trigger)).toBe('category=trigger')
-    })
-
-    it('should return endpoint category for extension', () => {
-      expect(getMarketplaceListCondition(PluginCategoryEnum.extension)).toBe('category=endpoint')
-    })
-
-    it('should return type condition for bundle', () => {
-      expect(getMarketplaceListCondition('bundle')).toBe('type=bundle')
-    })
-
-    it('should return empty string for all', () => {
-      expect(getMarketplaceListCondition('all')).toBe('')
-    })
-
-    it('should return empty string for unknown type', () => {
-      expect(getMarketplaceListCondition('unknown')).toBe('')
-    })
-  })
-
-  describe('getMarketplaceListFilterType', () => {
-    it('should return undefined for all', () => {
-      expect(getMarketplaceListFilterType(PLUGIN_TYPE_SEARCH_MAP.all)).toBeUndefined()
-    })
-
-    it('should return bundle for bundle', () => {
-      expect(getMarketplaceListFilterType(PLUGIN_TYPE_SEARCH_MAP.bundle)).toBe('bundle')
-    })
-
-    it('should return plugin for other categories', () => {
-      expect(getMarketplaceListFilterType(PLUGIN_TYPE_SEARCH_MAP.tool)).toBe('plugin')
-      expect(getMarketplaceListFilterType(PLUGIN_TYPE_SEARCH_MAP.model)).toBe('plugin')
-      expect(getMarketplaceListFilterType(PLUGIN_TYPE_SEARCH_MAP.agent)).toBe('plugin')
-    })
-  })
-})
-
-// ================================
-// useMarketplaceCollectionsAndPlugins Tests
-// ================================
-describe('useMarketplaceCollectionsAndPlugins', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  it('should return initial state correctly', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    expect(result.current.isLoading).toBe(false)
-    expect(result.current.isSuccess).toBe(false)
-    expect(result.current.queryMarketplaceCollectionsAndPlugins).toBeDefined()
-    expect(result.current.setMarketplaceCollections).toBeDefined()
-    expect(result.current.setMarketplaceCollectionPluginsMap).toBeDefined()
-  })
-
-  it('should provide queryMarketplaceCollectionsAndPlugins function', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    expect(typeof result.current.queryMarketplaceCollectionsAndPlugins).toBe('function')
-  })
-
-  it('should provide setMarketplaceCollections function', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    expect(typeof result.current.setMarketplaceCollections).toBe('function')
-  })
-
-  it('should provide setMarketplaceCollectionPluginsMap function', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    expect(typeof result.current.setMarketplaceCollectionPluginsMap).toBe('function')
-  })
-
-  it('should return marketplaceCollections from data or override', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    // Initial state
-    expect(result.current.marketplaceCollections).toBeUndefined()
-  })
-
-  it('should return marketplaceCollectionPluginsMap from data or override', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    // Initial state
-    expect(result.current.marketplaceCollectionPluginsMap).toBeUndefined()
-  })
-})
-
-// ================================
-// useMarketplacePluginsByCollectionId Tests
-// ================================
-describe('useMarketplacePluginsByCollectionId', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  it('should return initial state when collectionId is undefined', async () => {
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePluginsByCollectionId(undefined))
-
-    expect(result.current.plugins).toEqual([])
-    expect(result.current.isLoading).toBe(false)
-    expect(result.current.isSuccess).toBe(false)
-  })
-
-  it('should return isLoading false when collectionId is provided and query completes', async () => {
-    // The mock returns isFetching: false, isPending: false, so isLoading will be false
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePluginsByCollectionId('test-collection'))
-
-    // isLoading should be false since mock returns isFetching: false, isPending: false
-    expect(result.current.isLoading).toBe(false)
-  })
-
-  it('should accept query parameter', async () => {
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-    const { result } = renderHook(() =>
-      useMarketplacePluginsByCollectionId('test-collection', {
-        category: 'tool',
-        type: 'plugin',
-      }))
-
-    expect(result.current.plugins).toBeDefined()
-  })
-
-  it('should return plugins property from hook', async () => {
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePluginsByCollectionId('collection-1'))
-
-    // Hook should expose plugins property (may be array or fallback to empty array)
-    expect(result.current.plugins).toBeDefined()
-  })
-})
-
-// ================================
-// useMarketplacePlugins Tests
-// ================================
-describe('useMarketplacePlugins', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  it('should return initial state correctly', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(result.current.plugins).toBeUndefined()
-    expect(result.current.total).toBeUndefined()
-    expect(result.current.isLoading).toBe(false)
-    expect(result.current.isFetchingNextPage).toBe(false)
-    expect(result.current.hasNextPage).toBe(false)
-    expect(result.current.page).toBe(0)
-  })
-
-  it('should provide queryPlugins function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(typeof result.current.queryPlugins).toBe('function')
-  })
-
-  it('should provide queryPluginsWithDebounced function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(typeof result.current.queryPluginsWithDebounced).toBe('function')
-  })
-
-  it('should provide cancelQueryPluginsWithDebounced function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(typeof result.current.cancelQueryPluginsWithDebounced).toBe('function')
-  })
-
-  it('should provide resetPlugins function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(typeof result.current.resetPlugins).toBe('function')
-  })
-
-  it('should provide fetchNextPage function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(typeof result.current.fetchNextPage).toBe('function')
-  })
-
-  it('should normalize params with default pageSize', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // queryPlugins will normalize params internally
-    expect(result.current.queryPlugins).toBeDefined()
-  })
-
-  it('should handle queryPlugins call without errors', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Call queryPlugins
-    expect(() => {
-      result.current.queryPlugins({
-        query: 'test',
-        sort_by: 'install_count',
-        sort_order: 'DESC',
-        category: 'tool',
-        page_size: 20,
-      })
-    }).not.toThrow()
-  })
-
-  it('should handle queryPlugins with bundle type', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.queryPlugins({
-        query: 'test',
-        type: 'bundle',
-        page_size: 40,
-      })
-    }).not.toThrow()
-  })
-
-  it('should handle resetPlugins call', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.resetPlugins()
-    }).not.toThrow()
-  })
-
-  it('should handle queryPluginsWithDebounced call', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.queryPluginsWithDebounced({
-        query: 'debounced search',
-        category: 'all',
-      })
-    }).not.toThrow()
-  })
-
-  it('should handle cancelQueryPluginsWithDebounced call', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.cancelQueryPluginsWithDebounced()
-    }).not.toThrow()
-  })
-
-  it('should return correct page number', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Initially, page should be 0 when no query params
-    expect(result.current.page).toBe(0)
-  })
-
-  it('should handle queryPlugins with category all', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.queryPlugins({
-        query: 'test',
-        category: 'all',
-        sort_by: 'install_count',
-        sort_order: 'DESC',
-      })
-    }).not.toThrow()
-  })
-
-  it('should handle queryPlugins with tags', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.queryPlugins({
-        query: 'test',
-        tags: ['search', 'image'],
-        exclude: ['excluded-plugin'],
-      })
-    }).not.toThrow()
-  })
-
-  it('should handle queryPlugins with custom pageSize', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    expect(() => {
-      result.current.queryPlugins({
-        query: 'test',
-        page_size: 100,
-      })
-    }).not.toThrow()
-  })
-})
-
-// ================================
-// Hooks queryFn Coverage Tests
-// ================================
-describe('Hooks queryFn Coverage', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockInfiniteQueryData = undefined
-  })
-
-  it('should cover queryFn with pages data', async () => {
-    // Set mock data to have pages
-    mockInfiniteQueryData = {
-      pages: [
-        { plugins: [{ name: 'plugin1' }], total: 10, page: 1, page_size: 40 },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query to cover more code paths
-    result.current.queryPlugins({
-      query: 'test',
-      category: 'tool',
-    })
-
-    // With mockInfiniteQueryData set, plugin flatMap should be covered
-    expect(result.current).toBeDefined()
-  })
-
-  it('should expose page and total from infinite query data', async () => {
-    mockInfiniteQueryData = {
-      pages: [
-        { plugins: [{ name: 'plugin1' }, { name: 'plugin2' }], total: 20, page: 1, page_size: 40 },
-        { plugins: [{ name: 'plugin3' }], total: 20, page: 2, page_size: 40 },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // After setting query params, plugins should be computed
-    result.current.queryPlugins({
-      query: 'search',
-    })
-
-    // Hook returns page count based on mock data
-    expect(result.current.page).toBe(2)
-  })
-
-  it('should return undefined total when no query is set', async () => {
-    mockInfiniteQueryData = undefined
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // No query set, total should be undefined
-    expect(result.current.total).toBeUndefined()
-  })
-
-  it('should return total from first page when query is set and data exists', async () => {
-    mockInfiniteQueryData = {
-      pages: [
-        { plugins: [], total: 50, page: 1, page_size: 40 },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      query: 'test',
-    })
-
-    // After query, page should be computed from pages length
-    expect(result.current.page).toBe(1)
-  })
-
-  it('should cover queryFn for plugins type search', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query with plugin type
-    result.current.queryPlugins({
-      type: 'plugin',
-      query: 'search test',
-      category: 'model',
-      sort_by: 'version_updated_at',
-      sort_order: 'ASC',
-    })
-
-    expect(result.current).toBeDefined()
-  })
-
-  it('should cover queryFn for bundles type search', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query with bundle type
-    result.current.queryPlugins({
-      type: 'bundle',
-      query: 'bundle search',
-    })
-
-    expect(result.current).toBeDefined()
-  })
-
-  it('should handle empty pages array', async () => {
-    mockInfiniteQueryData = {
-      pages: [],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      query: 'test',
-    })
-
-    expect(result.current.page).toBe(0)
-  })
-
-  it('should handle API error in queryFn', async () => {
-    mockPostMarketplaceShouldFail = true
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Even when API fails, hook should still work
-    result.current.queryPlugins({
-      query: 'test that fails',
-    })
-
-    expect(result.current).toBeDefined()
-    mockPostMarketplaceShouldFail = false
-  })
-})
-
-// ================================
-// Advanced Hook Integration Tests
-// ================================
-describe('Advanced Hook Integration', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockInfiniteQueryData = undefined
-    mockPostMarketplaceShouldFail = false
-  })
-
-  it('should test useMarketplaceCollectionsAndPlugins with query call', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    // Call the query function
-    result.current.queryMarketplaceCollectionsAndPlugins({
-      condition: 'category=tool',
-      type: 'plugin',
-    })
-
-    expect(result.current.queryMarketplaceCollectionsAndPlugins).toBeDefined()
-  })
-
-  it('should test useMarketplaceCollectionsAndPlugins with empty query', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    // Call with undefined (converts to empty object)
-    result.current.queryMarketplaceCollectionsAndPlugins()
-
-    expect(result.current.queryMarketplaceCollectionsAndPlugins).toBeDefined()
-  })
-
-  it('should test useMarketplacePluginsByCollectionId with different params', async () => {
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-
-    // Test with various query params
-    const { result: result1 } = renderHook(() =>
-      useMarketplacePluginsByCollectionId('collection-1', {
-        category: 'tool',
-        type: 'plugin',
-        exclude: ['plugin-to-exclude'],
-      }))
-    expect(result1.current).toBeDefined()
-
-    const { result: result2 } = renderHook(() =>
-      useMarketplacePluginsByCollectionId('collection-2', {
-        type: 'bundle',
-      }))
-    expect(result2.current).toBeDefined()
-  })
-
-  it('should test useMarketplacePlugins with various parameters', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Test with all possible parameters
-    result.current.queryPlugins({
-      query: 'comprehensive test',
-      sort_by: 'install_count',
-      sort_order: 'DESC',
-      category: 'tool',
-      tags: ['tag1', 'tag2'],
-      exclude: ['excluded-plugin'],
-      type: 'plugin',
-      page_size: 50,
-    })
-
-    expect(result.current).toBeDefined()
-
-    // Test reset
-    result.current.resetPlugins()
-    expect(result.current.plugins).toBeUndefined()
-  })
-
-  it('should test debounced query function', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Test debounced query
-    result.current.queryPluginsWithDebounced({
-      query: 'debounced test',
-    })
-
-    // Cancel debounced query
-    result.current.cancelQueryPluginsWithDebounced()
-
-    expect(result.current).toBeDefined()
-  })
-})
-
-// ================================
-// Direct queryFn Coverage Tests
-// ================================
-describe('Direct queryFn Coverage', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockInfiniteQueryData = undefined
-    mockPostMarketplaceShouldFail = false
-    capturedInfiniteQueryFn = null
-    capturedQueryFn = null
-  })
-
-  it('should directly test useMarketplacePlugins queryFn execution', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-
-    // First render to capture queryFn
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query to set queryParams and enable the query
-    result.current.queryPlugins({
-      query: 'direct test',
-      category: 'tool',
-      sort_by: 'install_count',
-      sort_order: 'DESC',
-      page_size: 40,
-    })
-
-    // Now queryFn should be captured and enabled
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      // Call queryFn directly to cover internal logic
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-  })
-
-  it('should test queryFn with bundle type', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      type: 'bundle',
-      query: 'bundle test',
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 2, signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-  })
-
-  it('should test queryFn error handling', async () => {
-    mockPostMarketplaceShouldFail = true
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      query: 'test that will fail',
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      // This should trigger the catch block
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal })
-      expect(response).toBeDefined()
-      expect(response).toHaveProperty('plugins')
-    }
-
-    mockPostMarketplaceShouldFail = false
-  })
-
-  it('should test useMarketplaceCollectionsAndPlugins queryFn', async () => {
-    const { useMarketplaceCollectionsAndPlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplaceCollectionsAndPlugins())
-
-    // Trigger query to enable and capture queryFn
-    result.current.queryMarketplaceCollectionsAndPlugins({
-      condition: 'category=tool',
-    })
-
-    if (capturedQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedQueryFn({ signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-  })
-
-  it('should test queryFn with all category', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      category: 'all',
-      query: 'all category test',
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-  })
-
-  it('should test queryFn with tags and exclude', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      query: 'tags test',
-      tags: ['tag1', 'tag2'],
-      exclude: ['excluded1', 'excluded2'],
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-  })
-
-  it('should test useMarketplacePluginsByCollectionId queryFn coverage', async () => {
-    // Mock useQuery to capture queryFn from useMarketplacePluginsByCollectionId
-    const { useMarketplacePluginsByCollectionId } = await import('./hooks')
-
-    // Test with undefined collectionId - should return empty array in queryFn
-    const { result: result1 } = renderHook(() => useMarketplacePluginsByCollectionId(undefined))
-    expect(result1.current.plugins).toBeDefined()
-
-    // Test with valid collectionId - should call API in queryFn
-    const { result: result2 } = renderHook(() =>
-      useMarketplacePluginsByCollectionId('test-collection', { category: 'tool' }))
-    expect(result2.current).toBeDefined()
-  })
-
-  it('should test postMarketplace response with bundles', async () => {
-    // Temporarily modify mock response to return bundles
-    const originalBundles = [...mockPostMarketplaceResponse.data.bundles]
-    const originalPlugins = [...mockPostMarketplaceResponse.data.plugins]
-    mockPostMarketplaceResponse.data.bundles = [
-      { type: 'bundle', org: 'test', name: 'bundle1', tags: [] },
-    ]
-    mockPostMarketplaceResponse.data.plugins = []
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      type: 'bundle',
-      query: 'test bundles',
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal })
-      expect(response).toBeDefined()
-    }
-
-    // Restore original response
-    mockPostMarketplaceResponse.data.bundles = originalBundles
-    mockPostMarketplaceResponse.data.plugins = originalPlugins
-  })
-
-  it('should cover map callback with plugins data', async () => {
-    // Ensure API returns plugins
-    mockPostMarketplaceShouldFail = false
-    mockPostMarketplaceResponse.data.plugins = [
-      { type: 'plugin', org: 'test', name: 'plugin-for-map-1', tags: [] },
-      { type: 'plugin', org: 'test', name: 'plugin-for-map-2', tags: [] },
-    ]
-    mockPostMarketplaceResponse.data.total = 2
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Call queryPlugins to set queryParams (which triggers queryFn in our mock)
-    act(() => {
-      result.current.queryPlugins({
-        query: 'map coverage test',
-        category: 'tool',
-      })
-    })
-
-    // The queryFn is called by our mock when enabled is true
-    // Since we set queryParams, enabled should be true, and queryFn should be called
-    // with proper params, triggering the map callback
-    expect(result.current.queryPlugins).toBeDefined()
-  })
-
-  it('should test queryFn return structure', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({
-      query: 'structure test',
-      page_size: 20,
-    })
-
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 3, signal: controller.signal }) as {
-        plugins: unknown[]
-        total: number
-        page: number
-        page_size: number
-      }
-
-      // Verify the returned structure
-      expect(response).toHaveProperty('plugins')
-      expect(response).toHaveProperty('total')
-      expect(response).toHaveProperty('page')
-      expect(response).toHaveProperty('page_size')
-    }
-  })
-})
-
-// ================================
-// Line 198 flatMap Coverage Test
-// ================================
-describe('flatMap Coverage', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockPostMarketplaceShouldFail = false
-  })
-
-  it('should cover flatMap operation when data.pages exists', async () => {
-    // Set mock data with pages that have plugins
-    mockInfiniteQueryData = {
-      pages: [
-        {
-          plugins: [
-            { name: 'plugin1', type: 'plugin', org: 'test' },
-            { name: 'plugin2', type: 'plugin', org: 'test' },
-          ],
-          total: 5,
-          page: 1,
-          page_size: 40,
-        },
-        {
-          plugins: [
-            { name: 'plugin3', type: 'plugin', org: 'test' },
-          ],
-          total: 5,
-          page: 2,
-          page_size: 40,
-        },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query to set queryParams (hasQuery = true)
-    result.current.queryPlugins({
-      query: 'flatmap test',
-    })
-
-    // Hook should be defined
-    expect(result.current).toBeDefined()
-    // Query function should be triggered (coverage is the goal here)
-    expect(result.current.queryPlugins).toBeDefined()
-  })
-
-  it('should return undefined plugins when no query params', async () => {
-    mockInfiniteQueryData = undefined
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Don't trigger query, so hasQuery = false
-    expect(result.current.plugins).toBeUndefined()
-  })
-
-  it('should test hook with pages data for flatMap path', async () => {
-    mockInfiniteQueryData = {
-      pages: [
-        { plugins: [], total: 100, page: 1, page_size: 40 },
-        { plugins: [], total: 100, page: 2, page_size: 40 },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    result.current.queryPlugins({ query: 'total test' })
-
-    // Verify hook returns expected structure
-    expect(result.current.page).toBe(2) // pages.length
-    expect(result.current.queryPlugins).toBeDefined()
-  })
-
-  it('should handle API error and cover catch block', async () => {
-    mockPostMarketplaceShouldFail = true
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query that will fail
-    result.current.queryPlugins({
-      query: 'error test',
-      category: 'tool',
-    })
-
-    // Wait for queryFn to execute and handle error
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      try {
-        const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal }) as {
-          plugins: unknown[]
-          total: number
-          page: number
-          page_size: number
-        }
-        // When error is caught, should return fallback data
-        expect(response.plugins).toEqual([])
-        expect(response.total).toBe(0)
-      }
-      catch {
-        // This is expected when API fails
-      }
-    }
-
-    mockPostMarketplaceShouldFail = false
-  })
-
-  it('should test getNextPageParam directly', async () => {
-    const { useMarketplacePlugins } = await import('./hooks')
-    renderHook(() => useMarketplacePlugins())
-
-    // Test getNextPageParam function directly
-    if (capturedGetNextPageParam) {
-      // When there are more pages
-      const nextPage = capturedGetNextPageParam({ page: 1, page_size: 40, total: 100 })
-      expect(nextPage).toBe(2)
-
-      // When all data is loaded
-      const noMorePages = capturedGetNextPageParam({ page: 3, page_size: 40, total: 100 })
-      expect(noMorePages).toBeUndefined()
-
-      // Edge case: exactly at boundary
-      const atBoundary = capturedGetNextPageParam({ page: 2, page_size: 50, total: 100 })
-      expect(atBoundary).toBeUndefined()
-    }
-  })
-
-  it('should cover catch block by simulating API failure', async () => {
-    // Enable API failure mode
-    mockPostMarketplaceShouldFail = true
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result } = renderHook(() => useMarketplacePlugins())
-
-    // Set params to trigger the query
-    act(() => {
-      result.current.queryPlugins({
-        query: 'catch block test',
-        type: 'plugin',
-      })
-    })
-
-    // Directly invoke queryFn to trigger the catch block
-    if (capturedInfiniteQueryFn) {
-      const controller = new AbortController()
-      const response = await capturedInfiniteQueryFn({ pageParam: 1, signal: controller.signal }) as {
-        plugins: unknown[]
-        total: number
-        page: number
-        page_size: number
-      }
-      // Catch block should return fallback values
-      expect(response.plugins).toEqual([])
-      expect(response.total).toBe(0)
-      expect(response.page).toBe(1)
-    }
-
-    mockPostMarketplaceShouldFail = false
-  })
-
-  it('should cover flatMap when hasQuery and hasData are both true', async () => {
-    // Set mock data before rendering
-    mockInfiniteQueryData = {
-      pages: [
-        {
-          plugins: [{ name: 'test-plugin-1' }, { name: 'test-plugin-2' }],
-          total: 10,
-          page: 1,
-          page_size: 40,
-        },
-      ],
-    }
-
-    const { useMarketplacePlugins } = await import('./hooks')
-    const { result, rerender } = renderHook(() => useMarketplacePlugins())
-
-    // Trigger query to set queryParams
-    act(() => {
-      result.current.queryPlugins({
-        query: 'flatmap coverage test',
-      })
-    })
-
-    // Force rerender to pick up state changes
-    rerender()
-
-    // After rerender, hasQuery should be true
-    // The hook should compute plugins from pages.flatMap
-    expect(result.current).toBeDefined()
-  })
-})
-
-// ================================
-// Async Utils Tests
-// ================================
-
-// Narrow mock surface and avoid any in tests
-// Types are local to this spec to keep scope minimal
-
-type FnMock = ReturnType<typeof vi.fn>
-
-type MarketplaceClientMock = {
-  collectionPlugins: FnMock
-  collections: FnMock
-}
-
-describe('Async Utils', () => {
-  let marketplaceClientMock: MarketplaceClientMock
-
-  beforeAll(async () => {
-    const mod = await import('@/service/client')
-    marketplaceClientMock = mod.marketplaceClient as unknown as MarketplaceClientMock
-  })
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  afterEach(() => {
-    globalThis.fetch = originalFetch
-  })
-
-  describe('getMarketplacePluginsByCollectionId', () => {
-    it('should fetch plugins by collection id successfully', async () => {
-      const mockPlugins = [
-        { type: 'plugin', org: 'test', name: 'plugin1' },
-        { type: 'plugin', org: 'test', name: 'plugin2' },
-      ]
-
-      // Adjusted to our mocked marketplaceClient instead of fetch
-      marketplaceClientMock.collectionPlugins.mockResolvedValueOnce({
-        data: { plugins: mockPlugins },
-      })
-
-      const { getMarketplacePluginsByCollectionId } = await import('./utils')
-      const result = await getMarketplacePluginsByCollectionId('test-collection', {
-        category: 'tool',
-        exclude: ['excluded-plugin'],
-        type: 'plugin',
-      })
-
-      expect(marketplaceClientMock.collectionPlugins).toHaveBeenCalled()
-      expect(result).toHaveLength(2)
-    })
-
-    it('should handle fetch error and return empty array', async () => {
-      // Simulate error from client
-      marketplaceClientMock.collectionPlugins.mockRejectedValueOnce(new Error('Network error'))
-
-      const { getMarketplacePluginsByCollectionId } = await import('./utils')
-      const result = await getMarketplacePluginsByCollectionId('test-collection')
-
-      expect(result).toEqual([])
-    })
-
-    it('should pass abort signal when provided', async () => {
-      const mockPlugins = [{ type: 'plugins', org: 'test', name: 'plugin1' }]
-      // Our client mock receives the signal as second arg
-      marketplaceClientMock.collectionPlugins.mockResolvedValueOnce({
-        data: { plugins: mockPlugins },
-      })
-
-      const controller = new AbortController()
-      const { getMarketplacePluginsByCollectionId } = await import('./utils')
-      await getMarketplacePluginsByCollectionId('test-collection', {}, { signal: controller.signal })
-
-      expect(marketplaceClientMock.collectionPlugins).toHaveBeenCalled()
-      const call = marketplaceClientMock.collectionPlugins.mock.calls[0]
-      expect(call[1]).toMatchObject({ signal: controller.signal })
-    })
-  })
-
-  describe('getMarketplaceCollectionsAndPlugins', () => {
-    it('should fetch collections and plugins successfully', async () => {
-      const mockCollections = [
-        { name: 'collection1', label: {}, description: {}, rule: '', created_at: '', updated_at: '' },
-      ]
-      const mockPlugins = [{ type: 'plugins', org: 'test', name: 'plugin1' }]
-
-      // Simulate two-step client calls: collections then collectionPlugins
-      let stage = 0
-      marketplaceClientMock.collections.mockImplementationOnce(async () => {
-        stage = 1
-        return { data: { collections: mockCollections } }
-      })
-      marketplaceClientMock.collectionPlugins.mockImplementation(async () => {
-        if (stage === 1) {
-          return { data: { plugins: mockPlugins } }
-        }
-        return { data: { plugins: [] } }
-      })
-
-      const { getMarketplaceCollectionsAndPlugins } = await import('./utils')
-      const result = await getMarketplaceCollectionsAndPlugins({
-        condition: 'category=tool',
-        type: 'plugin',
-      })
-
-      expect(result.marketplaceCollections).toBeDefined()
-      expect(result.marketplaceCollectionPluginsMap).toBeDefined()
-    })
-
-    it('should handle fetch error and return empty data', async () => {
-      // Simulate client error
-      marketplaceClientMock.collections.mockRejectedValueOnce(new Error('Network error'))
-
-      const { getMarketplaceCollectionsAndPlugins } = await import('./utils')
-      const result = await getMarketplaceCollectionsAndPlugins()
-
-      expect(result.marketplaceCollections).toEqual([])
-      expect(result.marketplaceCollectionPluginsMap).toEqual({})
-    })
-
-    it('should append condition and type to URL when provided', async () => {
-      // Assert that the client was called with query containing condition/type
-      const { getMarketplaceCollectionsAndPlugins } = await import('./utils')
-      await getMarketplaceCollectionsAndPlugins({
-        condition: 'category=tool',
-        type: 'bundle',
-      })
-
-      expect(marketplaceClientMock.collections).toHaveBeenCalled()
-      const call = marketplaceClientMock.collections.mock.calls[0]
-      expect(call[0]).toMatchObject({ query: expect.objectContaining({ condition: 'category=tool', type: 'bundle' }) })
-    })
-  })
-})
-
-// ================================
-// useMarketplaceContainerScroll Tests
-// ================================
-describe('useMarketplaceContainerScroll', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  it('should attach scroll event listener to container', async () => {
-    const mockCallback = vi.fn()
-    const mockContainer = document.createElement('div')
-    mockContainer.id = 'marketplace-container'
-    document.body.appendChild(mockContainer)
-
-    const addEventListenerSpy = vi.spyOn(mockContainer, 'addEventListener')
-    const { useMarketplaceContainerScroll } = await import('./hooks')
-
-    const TestComponent = () => {
-      useMarketplaceContainerScroll(mockCallback)
-      return null
-    }
-
-    render(<TestComponent />)
-    expect(addEventListenerSpy).toHaveBeenCalledWith('scroll', expect.any(Function))
-    document.body.removeChild(mockContainer)
-  })
-
-  it('should call callback when scrolled to bottom', async () => {
-    const mockCallback = vi.fn()
-    const mockContainer = document.createElement('div')
-    mockContainer.id = 'scroll-test-container'
-    document.body.appendChild(mockContainer)
-
-    Object.defineProperty(mockContainer, 'scrollTop', { value: 900, writable: true })
-    Object.defineProperty(mockContainer, 'scrollHeight', { value: 1000, writable: true })
-    Object.defineProperty(mockContainer, 'clientHeight', { value: 100, writable: true })
-
-    const { useMarketplaceContainerScroll } = await import('./hooks')
-
-    const TestComponent = () => {
-      useMarketplaceContainerScroll(mockCallback, 'scroll-test-container')
-      return null
-    }
-
-    render(<TestComponent />)
-
-    const scrollEvent = new Event('scroll')
-    Object.defineProperty(scrollEvent, 'target', { value: mockContainer })
-    mockContainer.dispatchEvent(scrollEvent)
-
-    expect(mockCallback).toHaveBeenCalled()
-    document.body.removeChild(mockContainer)
-  })
-
-  it('should not call callback when scrollTop is 0', async () => {
-    const mockCallback = vi.fn()
-    const mockContainer = document.createElement('div')
-    mockContainer.id = 'scroll-test-container-2'
-    document.body.appendChild(mockContainer)
-
-    Object.defineProperty(mockContainer, 'scrollTop', { value: 0, writable: true })
-    Object.defineProperty(mockContainer, 'scrollHeight', { value: 1000, writable: true })
-    Object.defineProperty(mockContainer, 'clientHeight', { value: 100, writable: true })
-
-    const { useMarketplaceContainerScroll } = await import('./hooks')
-
-    const TestComponent = () => {
-      useMarketplaceContainerScroll(mockCallback, 'scroll-test-container-2')
-      return null
-    }
-
-    render(<TestComponent />)
-
-    const scrollEvent = new Event('scroll')
-    Object.defineProperty(scrollEvent, 'target', { value: mockContainer })
-    mockContainer.dispatchEvent(scrollEvent)
-
-    expect(mockCallback).not.toHaveBeenCalled()
-    document.body.removeChild(mockContainer)
-  })
-
-  it('should remove event listener on unmount', async () => {
-    const mockCallback = vi.fn()
-    const mockContainer = document.createElement('div')
-    mockContainer.id = 'scroll-unmount-container'
-    document.body.appendChild(mockContainer)
-
-    const removeEventListenerSpy = vi.spyOn(mockContainer, 'removeEventListener')
-    const { useMarketplaceContainerScroll } = await import('./hooks')
-
-    const TestComponent = () => {
-      useMarketplaceContainerScroll(mockCallback, 'scroll-unmount-container')
-      return null
-    }
-
-    const { unmount } = render(<TestComponent />)
-    unmount()
-
-    expect(removeEventListenerSpy).toHaveBeenCalledWith('scroll', expect.any(Function))
-    document.body.removeChild(mockContainer)
-  })
-})
-
-// ================================
-// Test Data Factory Tests
-// ================================
-describe('Test Data Factories', () => {
-  describe('createMockPlugin', () => {
-    it('should create plugin with default values', () => {
-      const plugin = createMockPlugin()
-
-      expect(plugin.type).toBe('plugin')
-      expect(plugin.org).toBe('test-org')
-      expect(plugin.version).toBe('1.0.0')
-      expect(plugin.verified).toBe(true)
-      expect(plugin.category).toBe(PluginCategoryEnum.tool)
-      expect(plugin.install_count).toBe(1000)
-    })
-
-    it('should allow overriding default values', () => {
-      const plugin = createMockPlugin({
-        name: 'custom-plugin',
-        org: 'custom-org',
-        version: '2.0.0',
-        install_count: 5000,
-      })
-
-      expect(plugin.name).toBe('custom-plugin')
-      expect(plugin.org).toBe('custom-org')
-      expect(plugin.version).toBe('2.0.0')
-      expect(plugin.install_count).toBe(5000)
-    })
-
-    it('should create bundle type plugin', () => {
-      const bundle = createMockPlugin({ type: 'bundle' })
-
-      expect(bundle.type).toBe('bundle')
-    })
-  })
-
-  describe('createMockPluginList', () => {
-    it('should create correct number of plugins', () => {
-      const plugins = createMockPluginList(5)
-
-      expect(plugins).toHaveLength(5)
-    })
-
-    it('should create plugins with unique names', () => {
-      const plugins = createMockPluginList(3)
-      const names = plugins.map(p => p.name)
-
-      expect(new Set(names).size).toBe(3)
-    })
-
-    it('should create plugins with decreasing install counts', () => {
-      const plugins = createMockPluginList(3)
-
-      expect(plugins[0].install_count).toBeGreaterThan(plugins[1].install_count)
-      expect(plugins[1].install_count).toBeGreaterThan(plugins[2].install_count)
-    })
-  })
-
-  describe('createMockCollection', () => {
-    it('should create collection with default values', () => {
-      const collection = createMockCollection()
-
-      expect(collection.name).toBe('test-collection')
-      expect(collection.label['en-US']).toBe('Test Collection')
-      expect(collection.searchable).toBe(true)
-    })
-
-    it('should allow overriding default values', () => {
-      const collection = createMockCollection({
-        name: 'custom-collection',
-        searchable: false,
-      })
-
-      expect(collection.name).toBe('custom-collection')
-      expect(collection.searchable).toBe(false)
-    })
+import { describe, it } from 'vitest'
+
+// The Marketplace index component is an async Server Component
+// that cannot be unit tested in jsdom. It is covered by integration tests.
+//
+// All sub-module tests have been moved to dedicated spec files:
+// - constants.spec.ts (DEFAULT_SORT, SCROLL_BOTTOM_THRESHOLD, PLUGIN_TYPE_SEARCH_MAP)
+// - utils.spec.ts (getPluginIconInMarketplace, getFormattedPlugin, getPluginLinkInMarketplace, etc.)
+// - hooks.spec.tsx (useMarketplaceCollectionsAndPlugins, useMarketplacePlugins, useMarketplaceContainerScroll)
+
+describe('Marketplace index', () => {
+  it('should be covered by dedicated sub-module specs', () => {
+    // Placeholder to document the split
   })
 })
