+++ web/app/components/plugins/plugin-page/filter-management/search-box.spec.tsx
@@ -8,22 +8,6 @@ vi.mock('react-i18next', () => ({
   }),
 }))
 
-vi.mock('@/app/components/base/input', () => ({
-  default: ({ value, onChange, placeholder }: {
-    value: string
-    onChange: (e: React.ChangeEvent<HTMLInputElement>) => void
-    placeholder: string
-    [key: string]: unknown
-  }) => (
-    <input
-      data-testid="search-input"
-      value={value}
-      onChange={onChange}
-      placeholder={placeholder}
-    />
-  ),
-}))
-
 describe('SearchBox', () => {
   let SearchBox: (typeof import('./search-box'))['default']
 
@@ -36,20 +20,20 @@ describe('SearchBox', () => {
   it('should render input with placeholder', () => {
     render(<SearchBox searchQuery="" onChange={vi.fn()} />)
 
-    expect(screen.getByTestId('search-input')).toHaveAttribute('placeholder', 'search')
+    expect(screen.getByRole('textbox')).toHaveAttribute('placeholder', 'search')
   })
 
   it('should display current search query', () => {
     render(<SearchBox searchQuery="test query" onChange={vi.fn()} />)
 
-    expect(screen.getByTestId('search-input')).toHaveValue('test query')
+    expect(screen.getByRole('textbox')).toHaveValue('test query')
   })
 
   it('should call onChange when input changes', () => {
     const mockOnChange = vi.fn()
     render(<SearchBox searchQuery="" onChange={mockOnChange} />)
 
-    fireEvent.change(screen.getByTestId('search-input'), { target: { value: 'new query' } })
+    fireEvent.change(screen.getByRole('textbox'), { target: { value: 'new query' } })
     expect(mockOnChange).toHaveBeenCalledWith('new query')
   })
 })
