+++ web/app/components/plugins/readme-panel/index.spec.tsx
@@ -3,7 +3,6 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { act, fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { PluginCategoryEnum, PluginSource } from '../types'
-import { BUILTIN_TOOLS_ARRAY } from './constants'
 import { ReadmeEntrance } from './entrance'
 import ReadmePanel from './index'
 import { ReadmeShowType, useReadmePanelStore } from './store'
@@ -115,289 +114,9 @@ const renderWithQueryClient = (ui: React.ReactElement) => {
   )
 }
 
-// ================================
-// Constants Tests
-// ================================
-describe('BUILTIN_TOOLS_ARRAY', () => {
-  it('should contain expected builtin tools', () => {
-    expect(BUILTIN_TOOLS_ARRAY).toContain('code')
-    expect(BUILTIN_TOOLS_ARRAY).toContain('audio')
-    expect(BUILTIN_TOOLS_ARRAY).toContain('time')
-    expect(BUILTIN_TOOLS_ARRAY).toContain('webscraper')
-  })
-
-  it('should have exactly 4 builtin tools', () => {
-    expect(BUILTIN_TOOLS_ARRAY).toHaveLength(4)
-  })
-})
-
-// ================================
-// Store Tests
-// ================================
-describe('useReadmePanelStore', () => {
-  describe('Initial State', () => {
-    it('should have undefined currentPluginDetail initially', () => {
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toBeUndefined()
-    })
-  })
-
-  describe('setCurrentPluginDetail', () => {
-    it('should set currentPluginDetail with detail and default showType', () => {
-      const mockDetail = createMockPluginDetail()
-      const { setCurrentPluginDetail } = useReadmePanelStore.getState()
-
-      act(() => {
-        setCurrentPluginDetail(mockDetail)
-      })
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toEqual({
-        detail: mockDetail,
-        showType: ReadmeShowType.drawer,
-      })
-    })
-
-    it('should set currentPluginDetail with custom showType', () => {
-      const mockDetail = createMockPluginDetail()
-      const { setCurrentPluginDetail } = useReadmePanelStore.getState()
-
-      act(() => {
-        setCurrentPluginDetail(mockDetail, ReadmeShowType.modal)
-      })
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toEqual({
-        detail: mockDetail,
-        showType: ReadmeShowType.modal,
-      })
-    })
-
-    it('should clear currentPluginDetail when called without arguments', () => {
-      const mockDetail = createMockPluginDetail()
-      const { setCurrentPluginDetail } = useReadmePanelStore.getState()
-
-      // First set a detail
-      act(() => {
-        setCurrentPluginDetail(mockDetail)
-      })
-
-      // Then clear it
-      act(() => {
-        setCurrentPluginDetail()
-      })
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toBeUndefined()
-    })
-
-    it('should clear currentPluginDetail when called with undefined', () => {
-      const mockDetail = createMockPluginDetail()
-      const { setCurrentPluginDetail } = useReadmePanelStore.getState()
-
-      // First set a detail
-      act(() => {
-        setCurrentPluginDetail(mockDetail)
-      })
-
-      // Then clear it with explicit undefined
-      act(() => {
-        setCurrentPluginDetail(undefined)
-      })
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toBeUndefined()
-    })
-  })
-
-  describe('ReadmeShowType enum', () => {
-    it('should have drawer and modal types', () => {
-      expect(ReadmeShowType.drawer).toBe('drawer')
-      expect(ReadmeShowType.modal).toBe('modal')
-    })
-  })
-})
-
-// ================================
-// ReadmeEntrance Component Tests
-// ================================
-describe('ReadmeEntrance', () => {
-  // ================================
-  // Rendering Tests
-  // ================================
-  describe('Rendering', () => {
-    it('should render the entrance button with full tip text', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(screen.getByRole('button')).toBeInTheDocument()
-      expect(screen.getByText('plugin.readmeInfo.needHelpCheckReadme')).toBeInTheDocument()
-    })
-
-    it('should render with short tip text when showShortTip is true', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} showShortTip />)
-
-      expect(screen.getByText('plugin.readmeInfo.title')).toBeInTheDocument()
-    })
-
-    it('should render divider when showShortTip is false', () => {
-      const mockDetail = createMockPluginDetail()
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} showShortTip={false} />)
-
-      expect(container.querySelector('.bg-divider-regular')).toBeInTheDocument()
-    })
-
-    it('should not render divider when showShortTip is true', () => {
-      const mockDetail = createMockPluginDetail()
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} showShortTip />)
-
-      expect(container.querySelector('.bg-divider-regular')).not.toBeInTheDocument()
-    })
-
-    it('should apply drawer mode padding class', () => {
-      const mockDetail = createMockPluginDetail()
-
-      const { container } = render(
-        <ReadmeEntrance pluginDetail={mockDetail} showType={ReadmeShowType.drawer} />,
-      )
-
-      expect(container.querySelector('.px-4')).toBeInTheDocument()
-    })
-
-    it('should apply custom className', () => {
-      const mockDetail = createMockPluginDetail()
-
-      const { container } = render(
-        <ReadmeEntrance pluginDetail={mockDetail} className="custom-class" />,
-      )
-
-      expect(container.querySelector('.custom-class')).toBeInTheDocument()
-    })
-  })
-
-  // ================================
-  // Conditional Rendering / Edge Cases
-  // ================================
-  describe('Conditional Rendering', () => {
-    it('should return null when pluginDetail is null/undefined', () => {
-      const { container } = render(<ReadmeEntrance pluginDetail={null as unknown as PluginDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should return null when plugin_unique_identifier is missing', () => {
-      const mockDetail = createMockPluginDetail({ plugin_unique_identifier: '' })
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should return null for builtin tool: code', () => {
-      const mockDetail = createMockPluginDetail({ id: 'code' })
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should return null for builtin tool: audio', () => {
-      const mockDetail = createMockPluginDetail({ id: 'audio' })
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should return null for builtin tool: time', () => {
-      const mockDetail = createMockPluginDetail({ id: 'time' })
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should return null for builtin tool: webscraper', () => {
-      const mockDetail = createMockPluginDetail({ id: 'webscraper' })
-
-      const { container } = render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(container.firstChild).toBeNull()
-    })
-
-    it('should render for non-builtin plugins', () => {
-      const mockDetail = createMockPluginDetail({ id: 'custom-plugin' })
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      expect(screen.getByRole('button')).toBeInTheDocument()
-    })
-  })
-
-  // ================================
-  // User Interactions / Event Handlers
-  // ================================
-  describe('User Interactions', () => {
-    it('should call setCurrentPluginDetail with drawer type when clicked', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      fireEvent.click(screen.getByRole('button'))
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toEqual({
-        detail: mockDetail,
-        showType: ReadmeShowType.drawer,
-      })
-    })
-
-    it('should call setCurrentPluginDetail with modal type when clicked', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} showType={ReadmeShowType.modal} />)
-
-      fireEvent.click(screen.getByRole('button'))
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail).toEqual({
-        detail: mockDetail,
-        showType: ReadmeShowType.modal,
-      })
-    })
-  })
-
-  // ================================
-  // Prop Variations
-  // ================================
-  describe('Prop Variations', () => {
-    it('should use default showType when not provided', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} />)
-
-      fireEvent.click(screen.getByRole('button'))
-
-      const { currentPluginDetail } = useReadmePanelStore.getState()
-      expect(currentPluginDetail?.showType).toBe(ReadmeShowType.drawer)
-    })
-
-    it('should handle modal showType correctly', () => {
-      const mockDetail = createMockPluginDetail()
-
-      render(<ReadmeEntrance pluginDetail={mockDetail} showType={ReadmeShowType.modal} />)
-
-      // Modal mode should not have px-4 class
-      const container = screen.getByRole('button').parentElement
-      expect(container).not.toHaveClass('px-4')
-    })
-  })
-})
+// Constants (BUILTIN_TOOLS_ARRAY) tests moved to constants.spec.ts
+// Store (useReadmePanelStore) tests moved to store.spec.ts
+// Entrance (ReadmeEntrance) tests moved to entrance.spec.tsx
 
 // ================================
 // ReadmePanel Component Tests
