+++ web/app/components/plugins/reference-setting-modal/index.spec.tsx
@@ -6,7 +6,6 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { PermissionType } from '@/app/components/plugins/types'
 import { AUTO_UPDATE_MODE, AUTO_UPDATE_STRATEGY } from './auto-update-setting/types'
 import ReferenceSettingModal from './index'
-import Label from './label'
 
 // ================================
 // Mock External Dependencies Only
@@ -156,153 +155,7 @@ describe('reference-setting-modal', () => {
     mockSystemFeatures.enable_marketplace = true
   })
 
-  // ============================================================
-  // Label Component Tests
-  // ============================================================
-  describe('Label (label.tsx)', () => {
-    describe('Rendering', () => {
-      it('should render label text', () => {
-        // Arrange & Act
-        render(<Label label="Test Label" />)
-
-        // Assert
-        expect(screen.getByText('Test Label')).toBeInTheDocument()
-      })
-
-      it('should render with label only when no description provided', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Simple Label" />)
-
-        // Assert
-        expect(screen.getByText('Simple Label')).toBeInTheDocument()
-        // Should have h-6 class when no description
-        expect(container.querySelector('.h-6')).toBeInTheDocument()
-      })
-
-      it('should render label and description when both provided', () => {
-        // Arrange & Act
-        render(<Label label="Label Text" description="Description Text" />)
-
-        // Assert
-        expect(screen.getByText('Label Text')).toBeInTheDocument()
-        expect(screen.getByText('Description Text')).toBeInTheDocument()
-      })
-
-      it('should apply h-4 class to label container when description is provided', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Label" description="Has description" />)
-
-        // Assert
-        expect(container.querySelector('.h-4')).toBeInTheDocument()
-      })
-
-      it('should not render description element when description is undefined', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Only Label" />)
-
-        // Assert
-        expect(container.querySelectorAll('.body-xs-regular')).toHaveLength(0)
-      })
-
-      it('should render description with correct styling', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Label" description="Styled Description" />)
-
-        // Assert
-        const descriptionElement = container.querySelector('.body-xs-regular')
-        expect(descriptionElement).toBeInTheDocument()
-        expect(descriptionElement).toHaveClass('mt-1', 'text-text-tertiary')
-      })
-    })
-
-    describe('Props Variations', () => {
-      it('should handle empty label string', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="" />)
-
-        // Assert - should render without crashing
-        expect(container.firstChild).toBeInTheDocument()
-      })
-
-      it('should handle empty description string', () => {
-        // Arrange & Act
-        render(<Label label="Label" description="" />)
-
-        // Assert - empty description still renders the description container
-        expect(screen.getByText('Label')).toBeInTheDocument()
-      })
-
-      it('should handle long label text', () => {
-        // Arrange
-        const longLabel = 'A'.repeat(200)
-
-        // Act
-        render(<Label label={longLabel} />)
-
-        // Assert
-        expect(screen.getByText(longLabel)).toBeInTheDocument()
-      })
-
-      it('should handle long description text', () => {
-        // Arrange
-        const longDescription = 'B'.repeat(500)
-
-        // Act
-        render(<Label label="Label" description={longDescription} />)
-
-        // Assert
-        expect(screen.getByText(longDescription)).toBeInTheDocument()
-      })
-
-      it('should handle special characters in label', () => {
-        // Arrange
-        const specialLabel = '<script>alert("xss")</script>'
-
-        // Act
-        render(<Label label={specialLabel} />)
-
-        // Assert - should be escaped
-        expect(screen.getByText(specialLabel)).toBeInTheDocument()
-      })
-
-      it('should handle special characters in description', () => {
-        // Arrange
-        const specialDescription = '!@#$%^&*()_+-=[]{}|;:,.<>?'
-
-        // Act
-        render(<Label label="Label" description={specialDescription} />)
-
-        // Assert
-        expect(screen.getByText(specialDescription)).toBeInTheDocument()
-      })
-    })
-
-    describe('Component Memoization', () => {
-      it('should be memoized with React.memo', () => {
-        // Assert
-        expect(Label).toBeDefined()
-        expect((Label as any).$$typeof?.toString()).toContain('Symbol')
-      })
-    })
-
-    describe('Styling', () => {
-      it('should apply system-sm-semibold class to label', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Styled Label" />)
-
-        // Assert
-        expect(container.querySelector('.system-sm-semibold')).toBeInTheDocument()
-      })
-
-      it('should apply text-text-secondary class to label', () => {
-        // Arrange & Act
-        const { container } = render(<Label label="Styled Label" />)
-
-        // Assert
-        expect(container.querySelector('.text-text-secondary')).toBeInTheDocument()
-      })
-    })
-  })
+  // Label component tests moved to label.spec.tsx
 
   // ============================================================
   // ReferenceSettingModal (PluginSettingModal) Component Tests
